requirements.txt
# Runtime dependencies
Django==5.2.1
djangorestframework==3.16.0
django-cors-headers==4.7.0
psycopg[binary]==3.2.6
Pillow==11.1.0
gunicorn==23.0.0
whitenoise==6.9.0
